+++ back-end/test-rules/projectsC.test.js
@@ -3,7 +3,7 @@
 */
 import { test, expect, describe, beforeAll } from '@jest/globals'
 
-import { dbAuth, FieldValue } from '@akauppi/firebase-jest-testing/firestoreReadOnly'
+import { dbAuth, FieldValue } from 'firebase-jest-testing/firestoreReadOnly'
 
 const anyDate = new Date();   // a non-server date